Add support for VS 2017
This commit links in the common files that the other VS* projects link
to, and also adds specific support for Visual Studio 2017. The only file
that is not linked in is the TfsHelper.Vs2012Base.cs file. This is
because Visual Studio 2017 is drastically different than Visual Studio
2012-2015 in terms of its installation and how you find information
about installed Visual Studio instances. Therefore, TfsHelper.Vs2017.cs
inherits directly from TfsHelperBase directly, and overrides the
required methods.

Visual Studio 2017 does not install many things into the Windows
Registry. There is guidance online that details how to determine the
installation path of an instance of Visual Studio 2017. (There can be
many instances of Visual Studio 2017 installed on any one machine now.)
This guidance has been used to determine the installation path, which in
turn is used to find the DLL containing the TFS Checkin Form.

The TFS Checkin Form is still in a DLL which is part of a Visual Studio
extension, but the extension now has a deterministic installation
location within the Visual Studio 2017 installation path.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>